feat: DHIS2-10509 moved query generation to separate class
