+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/association/IdentifiableObjectAssociations.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2004-2004-2020, University of Oslo
+ * Copyright (c) 2004-2021, University of Oslo
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
